test(uploads): move the upload-naming tests to their own file
