+++ tests/integration/links/test_session_cover_repository.py
@@ -1,5 +1,3 @@
-import re
-
 from django.core.files.uploadedfile import SimpleUploadedFile
 
 from ludamus.links.db.django.repositories import SessionRepository, delete_stored_file
@@ -32,30 +30,3 @@ class TestSessionRepositoryCoverUpdate:
         session.refresh_from_db()
         assert session.cover_image.name != old_name
         assert not storage.exists(old_name)
-
-
-class TestUniqueUploadTo:
-    def test_same_filename_uploaded_twice_gets_distinct_names(self, agenda_item):
-        session = agenda_item.session
-        stored = r"sessions/[0-9a-f]{32}\.png"
-
-        names = []
-        for _ in range(2):
-            session.cover_image = SimpleUploadedFile(
-                "image.PNG", PNG_BYTES, content_type="image/png"
-            )
-            session.save()
-            names.append(session.cover_image.name)
-
-        assert names[0] != names[1]
-        assert all(re.fullmatch(stored, name) for name in names)
-
-    def test_unlisted_suffix_is_dropped(self, agenda_item):
-        session = agenda_item.session
-
-        session.cover_image = SimpleUploadedFile(
-            "image.html", PNG_BYTES, content_type="image/png"
-        )
-        session.save()
-
-        assert re.fullmatch(r"sessions/[0-9a-f]{32}", session.cover_image.name)
